Project updated

+++ ProjetoFinalAliare/Frm_ConsultarCurso.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
+using System.Data.Entity.Infrastructure;
 using System.Drawing;
 using System.Linq;
 using System.Text;
@@ -16,6 +17,8 @@ namespace ProjetoFinalAliare
         public Frm_ConsultarCurso()
         {
             InitializeComponent();
+            Btn_Editar.Enabled = false;
+            Btn_Deletar.Enabled = false;
         }
 
         private void Btn_Voltar_Click(object sender, EventArgs e)
@@ -32,30 +35,54 @@ namespace ProjetoFinalAliare
         private void Btn_Consulta_Click(object sender, EventArgs e)
         {
             LimparTextBoxes();
-            dataGridView1.DataSource = CursoController.lerCursos();
-            
+            dataGridView1.DataSource = CursoController.SelecionarCursos();
+            Btn_Editar.Enabled = true;
+            Btn_Deletar.Enabled = true;
+
         }
 
         private void Btn_Editar_Click(object sender, EventArgs e)
         {
-            var form = new Frm_EditarCurso(Txb_IdCurso.Text);
-            form.ShowDialog();
+            if (Txb_IdCurso.Text == "")
+            {
+                MensagemDeSelecao();
+            }
+            else
+            {
+                var form = new Frm_EditarCurso(Txb_IdCurso.Text);
+                form.ShowDialog();
+            }
         }
        
         private void Btn_Deletar_Click(object sender, EventArgs e)
         {
-            var idCurso = int.Parse(Txb_IdCurso.Text);
-            //Criar um MessageBox com os botões Sim e Não e deixar o botão 2(Não) selecionado por padrão e comparar o botão apertado
-            if (DialogResult.Yes == MessageBox.Show("Tem certeza que deseja apagar o registro?", "Confirmação", MessageBoxButtons.YesNo, MessageBoxIcon.Question, MessageBoxDefaultButton.Button2))
+            try
             {
-                //Rotina de exclusão
-                CursoController.DeleteCurso(idCurso);
-                dataGridView1.DataSource = CursoController.lerCursos();
+                if (Txb_IdCurso.Text == "")
+                {
+                    MensagemDeSelecao();
+                }
+                else
+                {
+                    var idCurso = int.Parse(Txb_IdCurso.Text);
+                    //Criar um MessageBox com os botões Sim e Não e deixar o botão 2(Não) selecionado por padrão e comparar o botão apertado
+                    if (DialogResult.Yes == MessageBox.Show("Tem certeza que deseja apagar o registro?", "Confirmação", MessageBoxButtons.YesNo, MessageBoxIcon.Question, MessageBoxDefaultButton.Button2))
+                    {
+                        //Rotina de exclusão
+                        CursoController.DeleteCurso(idCurso);
+                        dataGridView1.DataSource = CursoController.SelecionarCursos();
 
-                //Confirmando exclusão para o usuário
-                MessageBox.Show("Registro apagado com sucesso", "Sucesso", MessageBoxButtons.OK, MessageBoxIcon.Information);
-
+                        //Confirmando exclusão para o usuário
+                        MessageBox.Show("Registro apagado com sucesso", "Sucesso", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    }
+                }
             }
+            catch (DbUpdateException)
+            {
+                MessageBox.Show("O curso não pode ser excluído. Favor verificar sem existe algum aluno matriculado no curso.", "Aliare", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                
+            }
+            
         }
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
@@ -77,6 +104,11 @@ namespace ProjetoFinalAliare
         {
             Txb_IdCurso.Text = null;
             Txb_NomeCurso.Text = null;
+        }
+
+        private void MensagemDeSelecao()
+        {
+            MessageBox.Show("Por favor, selecione um aluno.", "Selecionar aluno", MessageBoxButtons.OK, MessageBoxIcon.Warning);
         }
     }
 }
